@@ -1,25 +1,16 @@
 package com.futurerenewables;
 
 import android.app.Application;
+import android.util.Log;
 
+import com.facebook.react.PackageList;
+import com.facebook.hermes.reactexecutor.HermesExecutorFactory;
+import com.facebook.react.bridge.JavaScriptExecutorFactory;
 import com.facebook.react.ReactApplication;
-import com.hieuvp.fingerprint.ReactNativeFingerprintScannerPackage;
-import org.wonday.orientation.OrientationPackage;
-import org.reactnative.camera.RNCameraPackage;
-import com.horcrux.svg.SvgPackage;
-import com.imagepicker.ImagePickerPackage;
-import com.robinpowered.react.Intercom.IntercomPackage;
-import com.amazonaws.RNAWSCognitoPackage;
-import com.BV.LinearGradient.LinearGradientPackage;
-import com.airbnb.android.react.maps.MapsPackage;
-import com.brentvatne.react.ReactVideoPackage;
-import com.swmansion.gesturehandler.react.RNGestureHandlerPackage;
 import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
-import com.facebook.react.shell.MainReactPackage;
 import com.facebook.soloader.SoLoader;
 
-import java.util.Arrays;
 import java.util.List;
 
 public class MainApplication extends Application implements ReactApplication {
@@ -32,20 +23,11 @@ public class MainApplication extends Application implements ReactApplication {
 
     @Override
     protected List<ReactPackage> getPackages() {
-      return Arrays.<ReactPackage>asList(
-          new MainReactPackage(),
-            new ReactNativeFingerprintScannerPackage(),
-            new OrientationPackage(),
-            new RNCameraPackage(),
-            new SvgPackage(),
-            new ImagePickerPackage(),
-            new IntercomPackage(),
-            new RNAWSCognitoPackage(),
-            new LinearGradientPackage(),
-            new MapsPackage(),
-            new ReactVideoPackage(),
-            new RNGestureHandlerPackage()
-      );
+      @SuppressWarnings("UnnecessaryLocalVariable")
+      List<ReactPackage> packages = new PackageList(this).getPackages();
+      // Packages that cannot be autolinked yet can be added manually here, for example:
+      // packages.add(new MyReactNativePackage());
+      return packages;
     }
 
     @Override
