@@ -118,13 +118,7 @@ export function formatFullDate(newDate) {
   }
   oldDateStripped = f;
   let final;
-  if (f.length < 2) {
-    final = f;
-  } else if (f.length < 4) {
-    final = `${f.slice(0, 2)}${'/'}${f.slice(2, 4)}`;
-  } else {
-    final = `${f.slice(0, 2)}${'/'}${f.slice(2, 4)}${'/'}${f.slice(4, 8)}`;
-  }
+  if (f.length < 2) { final = f; } else if (f.length < 4) { final = `${f.slice(0, 2)}${'/'}${f.slice(2, 4)}`; } else { final = `${f.slice(0, 2)}${'/'}${f.slice(2, 4)}${'/'}${f.slice(4, 8)}`; }
   return final;
 }
 
